refactor: extract requestListener to its own file
