+++ __tests__/src/lib/__request_fixtures.js
@@ -1,0 +1,7 @@
+import { A_URL } from './__list_fixtures.js';
+
+export const RANDOM_REQUEST_DETAILS = { url: A_URL };
+export const BLOCKING_RESPONSE = {
+  CANCELLED: { cancel: true },
+  NOT_CANCELLED: { cancel: false },
+};
