+++ src/lib/request_blocker.js
@@ -1,24 +1,14 @@
-import RequestMatcher from './request_matcher.js';
+import requestListener from './request_listener.js';
 
-function requestBlockerListener(requestDetails) {
-  const url = requestDetails.url;
-  const isUrlDenied = RequestMatcher.isDenied(url);
-  const shouldBlock = { cancel: isUrlDenied };
-
-  if (isUrlDenied) {
-    console.log(`Blocking request for: ${url}. Nice try!`);
-  } else {
-    console.log(`URL is good to go: ${url}.`);
-  }
-  return shouldBlock;
-}
-
+/**
+ * @see https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/webRequest/onBeforeRequest
+ */
 export default class RequestBlocker {
   static startMonitoring() {
     const BLOCKING_FLAG = 'blocking';
 
     browser.webRequest.onBeforeRequest.addListener(
-      requestBlockerListener,
+      requestListener,
       { urls: ['<all_urls>'] },
       [BLOCKING_FLAG]
     );
